backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.10
Django==5.1.2
django-cors-headers==4.6.0
django-health-check==3.18.3
djangorestframework==3.15.2
idna==3.10
Khayyam==3.0.17
kombu==5.5.0
minio==7.2.12
prompt_toolkit==3.0.50
PyMySQL==1.1.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
